move reference_docs_generator to a separate dictionary
