+++ docs_generator/__init__.py
@@ -1,0 +1,1 @@
+from docs_generator.reference_docs_generator import ReferenceDocsGenerator
