+++ docs_generator/reference_docs_generator_test.py
@@ -1,9 +1,9 @@
 from pathlib import Path
 
-from src.conftest import BarCommand, FooCommand
+from conftest import BarCommand, FooCommand
+from docs_generator import ReferenceDocsGenerator
 from src.cli.cli_app import CLIApp
 from src.cli.command import Command
-from src.cli.reference_docs_generator import ReferenceDocsGenerator
 
 
 def test_generating_markdown_for_commands(
