+++ src/cli/__init__.py
@@ -4,4 +4,3 @@ from .argument_default_value_from_config_provider import (
 from .argument_parser_facade import ArgumentParserFacade
 from .cli_app import CLIApp
 from .command import Command
-from .reference_docs_generator import ReferenceDocsGenerator
