+++ src/cli/argument_default_value_from_config_provider_test.py
@@ -1,9 +1,9 @@
 from pytest_mock import MockerFixture
 
-from src.conftest import FooCommand
 from src.cli.argument_default_value_from_config_provider import (
     ArgumentDefaultValueFromConfigProvider,
 )
+from conftest import FooCommand
 from src.utils import Project
